Add files via upload
Messrs. Baker, Dyer, Farmer, Glover and Hosier are seated around a circular table, playing cards. Each
person has the namesake of the profession of one of the others. The dyer is seated two places to the left of
Mr. Hosier. The baker sits two places to Mr. Baker’s right. The farmer is seated to the left of Mr. Farmer.
Mr. Dyer is on the glover’s right. 
What is the name of the Dyer?